Revert "Integrate Kanban board into home page"
